refactor(cloudmonitoring): move MetricExporter to cloudotel
